+++ BitMap.h
@@ -20,8 +20,8 @@ private:
 
 public:
 	void Init(HDC hdc, char* FileName);
-	void Draw(HDC hdc, int x, int y, int IMG_WSize, int IMG_HSize);
-	void DrawTransparent(HDC hdc, int x, int y, int IMG_WSize, int IMG_HSize); //투명하게 출력(투명도)
+	void Draw(HDC hdc, float x, float y, float IMG_WSize, float IMG_HSize);
+	void DrawTransparent(HDC hdc, float x, float y, float IMG_WSize, float IMG_HSize); //투명하게 출력(투명도)
 	inline SIZE GetSize() { return m_Size; }
 	BitMap();
 	~BitMap();
